Splitting sorters into seperate classes.
