+++ src/services/sorters/BggRatingSorter.ts
@@ -1,0 +1,14 @@
+import { Sorter } from "./Sorter";
+import { GameInfoPlus, GameInfo } from "../../models/GameInfo";
+
+export class BggRatingSorter implements Sorter {
+    sort(collection: GameInfoPlus[]): GameInfoPlus[] {
+        return [...collection.sort(this.averageRatingSorter)];
+    }
+    private averageRatingSorter(a: GameInfo, b: GameInfo): number {
+        return b.averagerating - a.averagerating;
+    }
+}
+
+
+
