+++ src/services/sorters/NameSorter.ts
@@ -1,0 +1,18 @@
+import { Sorter } from "./Sorter";
+import { GameInfoPlus, GameInfo } from "../../models/GameInfo";
+
+export class NameSorter implements Sorter {
+    sort(collection: GameInfoPlus[]): GameInfoPlus[] {
+        return [...collection.sort(this.nameSorter)];
+    }
+    private nameSorter(a: GameInfo, b: GameInfo): number {
+        return a.name.localeCompare(b.name);
+    }
+
+}
+
+
+
+
+
+
