+++ src/services/sorters/Sorter.ts
@@ -1,0 +1,8 @@
+import { GameInfoPlus } from "../../models/GameInfo";
+
+/**
+ * A soter takes a collection and returns that collection sorted.
+ */
+export interface Sorter {
+    sort(collection: GameInfoPlus[]): GameInfoPlus[];
+}
